requirements.txt
wheel==0.23.0
click
matplotlib
numpy
requests
pillow
